requirements.txt
chalkpy[runtime]
google-generativeai
